+++ previous-versions/commerce/src/codatcommerce/models/shared/companyinfo.py
@@ -32,7 +32,7 @@ class CompanyInfoWeblink:
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class CompanyInfo:
-    r"""In the Codat system, company information includes standard commercial details about
+    r"""In the Codat system, company profile includes standard commercial details about
     a linked company, such as their address, phone number, and company registration.
 
     Explore our [data coverage](https://knowledge.codat.io/supported-features/commerce?view=tab-by-data-type&dataType=companyInfo) for this data type.