+++ previous-versions/commerce/RELEASES.md
@@ -637,3 +637,13 @@ Based on:
 - [python v0.35.0] previous-versions/commerce
 ### Releases
 - [PyPI v0.35.0] https://pypi.org/project/codat-commerce/0.35.0 - previous-versions/commerce
+
+## 2023-10-30 01:27:27
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.35.1] previous-versions/commerce
+### Releases
+- [PyPI v0.35.1] https://pypi.org/project/codat-commerce/0.35.1 - previous-versions/commerce